Upstream version 4.03.00

+++ irandstream/irandstream3.f
@@ -1,4 +1,4 @@
-inline IRandStream::IRandStream(int min, int max, int seed)
+inline IRandStream::IRandStream(int min, int max, long seed)
 :
     Randbuffer(min, max, seed),
     std::istream(this)
